% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpe.R
\name{beta_hat}
\alias{beta_hat}
\title{beta_hat}
\usage{
beta_hat(x, h, X, Y, p)
}
\arguments{
\item{x}{the argument}

\item{h}{the bandwidth}

\item{X}{the X of multivariate sample point (X,Y)}

\item{Y}{the X of multivariate sample point (X,Y)}

\item{p}{the order of polynomial}
}
\description{
helper function(calculate the beta_hat of local polynomial estimation.)
}
\examples{
\dontrun{
beta_hat(x,h,X,Y,3)
} 
}
