% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpe.R, R/op_bd.R
\name{CV}
\alias{CV}
\title{CV}
\usage{
CV(h)

CV(h)
}
\arguments{
\item{h}{the bandwidth}
}
\value{
The function value
}
\description{
helper function(calculate the optimal bandwidth)

helper function(f1-f2)
}
\examples{
\dontrun{
CV(h)
} 
\dontrun{
CV(h)
} 
}
