% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpe.R
\name{f}
\alias{f}
\title{f}
\usage{
f(h, X, Y)
}
\arguments{
\item{h}{the bandwidth}

\item{X}{the X of multivariate sample point (X,Y)}

\item{Y}{the X of multivariate sample point (X,Y)}
}
\description{
helper function(plotting the figure of beta_hat(x))
}
\examples{
\dontrun{
f(h,X,Y)
} 
}
