% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_bd.R
\name{f1}
\alias{f1}
\title{f1}
\usage{
f1(h)
}
\arguments{
\item{h}{the bandwidth}
}
\value{
The function value
}
\description{
helper function(the first_term in the cross-validation score function.)
}
\examples{
\dontrun{
f1(h)
} 
}
