% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_bd.R
\name{f2}
\alias{f2}
\title{f2}
\usage{
f2(h)
}
\arguments{
\item{h}{the bandwidth}
}
\value{
The function value
}
\description{
helper function(the second term in the cross-validation score function.)
}
\examples{
\dontrun{
f2(h)
} 
}
