% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_bd.R
\name{fh_hat}
\alias{fh_hat}
\title{fh_hat}
\usage{
fh_hat(x, h)
}
\arguments{
\item{x}{the argument}

\item{h}{the bandwidth}
}
\value{
The function value
}
\description{
$fh_hat(x)$
}
\examples{
\dontrun{
fh_hat(h,x)
} 
}
