% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpe.R, R/op_bd.R
\name{Kf}
\alias{Kf}
\title{Kf}
\usage{
Kf(x, h, kernel)

Kf(x, h, kernel)
}
\arguments{
\item{x}{the argument}

\item{h}{the bandwidth}

\item{kernel}{the kernel function}
}
\value{
The function value

The function value
}
\description{
helper function(choose the kernel function)

helper function(choose the kernel function)
}
\examples{
\dontrun{
Kf(x,h,"Triangle")
} 
\dontrun{
Kf(x,h,"Triangle")
} 
}
