% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpe.R
\name{lpe}
\alias{lpe}
\title{local polynomial estimation}
\usage{
lpe(X, Y, p, kernel)
}
\arguments{
\item{X}{the X of multivariate sample point (X,Y)}

\item{Y}{the X of multivariate sample point (X,Y)}

\item{p}{the order of polynomial}

\item{kernel}{the type of kernel function(6 types of kernel functions in this function)}
}
\value{
three plots of the polynomial estimation(the Plot of X and estimated function,the plot of beta_hat with optimal bandwidth,plots of beta_hat with the bandwidths of the interval (0,1))
}
\description{
main function(Use R to do local polynomial estimation and make a plot of the beta_hat.)
}
\examples{
\dontrun{
lpe(X,Y,P,"Triangle")
} 
}
