% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_bd.R
\name{op_bd}
\alias{op_bd}
\title{optimal bandwidth}
\usage{
op_bd(X, kernel)
}
\arguments{
\item{X}{several samples}

\item{kernel}{the type of kernel function(6 types of kernel functions in this function)}
}
\value{
two plots (the Plot of CV_score function,the plot of f_hat(x) with optimal bandwidth)
}
\description{
main function(get the optimal bandwidth and make a plot of f_hat(x) with optimal bandwidth.)
}
\examples{
\dontrun{
X<-mtcars$wt
op_bd(X,"Triangle")
} 
}
